Raised abstraction through a portion of query-serialization. Working through a few remaining tests
