+++ src/addons/meta/api.c
@@ -44,249 +44,6 @@ bool flecs_type_is_number(
     }
 }
 
-void flecs_json_ser_bool(bool value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    if (value) {
-        ecs_strbuf_appendlit(str, "true");
-    } else {
-        ecs_strbuf_appendlit(str, "false");
-    }
-}
-
-void flecs_json_ser_char_expression(char value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    if (value) {
-        char chbuf[3];
-        flecs_chresc(chbuf, value, '"');
-        ecs_strbuf_appendch(str, '"');
-        ecs_strbuf_appendstr(str, chbuf);
-        ecs_strbuf_appendch(str, '"');
-    } else {
-        ecs_strbuf_appendch(str, '0');
-    }
-}
-void flecs_json_ser_char(char value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    if (value) {
-        char chbuf[3];
-        flecs_chresc(chbuf, value, '"');
-        ecs_strbuf_appendstr(str, chbuf);
-    } else {
-        ecs_strbuf_appendch(str, '0');
-    }
-}
-
-void flecs_json_ser_byte(const uint8_t value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_appendint(str, flecs_uto(int64_t, value));
-}
-
-void flecs_json_ser_u8(const uint8_t value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_appendint(str, flecs_uto(int64_t, value));
-}
-
-void flecs_json_ser_u16(const uint16_t value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_appendint(str, flecs_uto(int64_t, value));
-}
-
-void flecs_json_ser_u32(const uint32_t value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_appendint(str, flecs_uto(int64_t, value));
-}
-
-void flecs_json_ser_u64(const uint64_t value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_append(str, "%llu", value);
-}
-
-void flecs_json_ser_i8(const int8_t value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_appendint(str, flecs_ito(int64_t, value));
-}
-
-void flecs_json_ser_i16(const int16_t value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_appendint(str, flecs_ito(int64_t, value));
-}
-
-void flecs_json_ser_i32(const int32_t value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_appendint(str, flecs_ito(int64_t, value));
-}
-
-void flecs_json_ser_i64(const int64_t value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_appendint(str, value);
-}
-
-void flecs_json_ser_f32(const float value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_appendflt(str, (double)value, 0);
-}
-
-void flecs_json_ser_f32_expression(const ecs_f32_t value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_appendflt(str, (ecs_f64_t)value, '"');
-}
-
-void flecs_json_ser_f64(const double value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_appendflt(str, value, 0);
-}
-
-void flecs_json_ser_f64_expression(const ecs_f64_t value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_appendflt(str, value, '"');
-}
-
-void flecs_json_ser_iptr(const intptr_t value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_appendint(str, flecs_ito(int64_t, value));
-}
-
-void flecs_json_ser_uptr(const uintptr_t value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    ecs_strbuf_append(str, "%u", value);
-}
-
-void flecs_json_ser_string_expression(const char* value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    if (value) {
-        ecs_size_t length = flecs_stresc(NULL, 0, '"', value);
-        if (length == ecs_os_strlen(value)) {
-            ecs_strbuf_appendch(str, '"');
-            ecs_strbuf_appendstrn(str, value, length);
-            ecs_strbuf_appendch(str, '"');
-        } else {
-            char *out = ecs_os_malloc(length + 3);
-            flecs_stresc(out + 1, length, '"', value);
-            out[0] = '"';
-            out[length + 1] = '"';
-            out[length + 2] = '\0';
-            ecs_strbuf_appendstr(str, out);
-            ecs_os_free(out);
-        }
-    } else {
-        ecs_strbuf_appendlit(str, "null");
-    }
-}
-
-void flecs_json_ser_string(const char* value, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    if (value) {
-        ecs_strbuf_appendstr(str, value);
-    } else {
-        ecs_strbuf_appendlit(str, "null");
-    }
-}
-
-void flecs_json_ser_entity(const ecs_world_t* world, ecs_entity_t e, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    if (!e) {
-        ecs_strbuf_appendlit(str, "#0");
-    } else {
-        ecs_get_path_w_sep_buf(world, 0, e, ".", NULL, str, false);
-    }
-}
-
-void flecs_json_ser_id(const ecs_world_t* world, ecs_id_t id, void* user_data) {
-    ecs_strbuf_t* str = (ecs_strbuf_t*)user_data;
-    if (!id) {
-        ecs_strbuf_appendlit(str, "#0");
-    } else {
-        ecs_id_str_buf(world, id, str);
-    }
-}
-
-/* Serialize a primitive value */
-int flecs_expr_ser_primitive(
-    const ecs_world_t *world,
-    ecs_primitive_kind_t kind,
-    ecs_visitor_desc_t *visitor_desc,
-    const void *base) 
-{
-    switch(kind) {
-    case EcsBool:
-        if (visitor_desc->visit_bool)
-            visitor_desc->visit_bool(*(const bool*)base, visitor_desc->user_data);
-        break;
-    case EcsChar:
-        if (visitor_desc->visit_char)
-            visitor_desc->visit_char(*(const char*)base, visitor_desc->user_data);
-        break;
-    case EcsByte:
-        if (visitor_desc->visit_byte)
-            visitor_desc->visit_byte(*(const uint8_t*)base, visitor_desc->user_data);
-        break;
-    case EcsU8:
-        if (visitor_desc->visit_u8)
-            visitor_desc->visit_u8(*(const uint8_t*)base, visitor_desc->user_data);
-        break;
-    case EcsU16:
-        if (visitor_desc->visit_u16)
-            visitor_desc->visit_u16(*(const uint16_t*)base, visitor_desc->user_data);
-        break;
-    case EcsU32:
-        if (visitor_desc->visit_u32)
-            visitor_desc->visit_u32(*(const uint32_t*)base, visitor_desc->user_data);
-        break;
-    case EcsU64:
-        if (visitor_desc->visit_u64)
-            visitor_desc->visit_u64(*(const uint64_t*)base, visitor_desc->user_data);
-        break;
-    case EcsI8:
-        if (visitor_desc->visit_i8)
-            visitor_desc->visit_i8(*(const int8_t*)base, visitor_desc->user_data);
-        break;
-    case EcsI16:
-        if (visitor_desc->visit_i16)
-            visitor_desc->visit_i16(*(const int16_t*)base, visitor_desc->user_data);
-        break;
-    case EcsI32:
-        if (visitor_desc->visit_i32)
-            visitor_desc->visit_i32(*(const int32_t*)base, visitor_desc->user_data);
-        break;
-    case EcsI64:
-        if (visitor_desc->visit_i64)
-            visitor_desc->visit_i64(*(const int64_t*)base, visitor_desc->user_data);
-        break;
-    case EcsF32:
-        if (visitor_desc->visit_f32)
-            visitor_desc->visit_f32(*(const float*)base, visitor_desc->user_data);
-        break;
-    case EcsF64:
-        if (visitor_desc->visit_f64)
-            visitor_desc->visit_f64(*(const double*)base, visitor_desc->user_data);
-        break;
-    case EcsIPtr:
-        if (visitor_desc->visit_iptr)
-            visitor_desc->visit_iptr(*(const intptr_t*)base, visitor_desc->user_data);
-        break;
-    case EcsUPtr:
-        if (visitor_desc->visit_uptr)
-            visitor_desc->visit_uptr(*(const uintptr_t*)base, visitor_desc->user_data);
-        break;
-    case EcsString:
-        if (visitor_desc->visit_string)
-            visitor_desc->visit_string(*ECS_CONST_CAST(const char**, base), visitor_desc->user_data);
-        break;
-    case EcsEntity:
-        if (visitor_desc->visit_entity)
-            visitor_desc->visit_entity(world, *(const ecs_entity_t*)base, visitor_desc->user_data);
-        break;
-    case EcsId:
-        if (visitor_desc->visit_id)
-            visitor_desc->visit_id(world, *(const ecs_id_t*)base, visitor_desc->user_data);
-        break;
-    default:
-        ecs_err("invalid primitive kind");
-        return -1;
-    }
-
-    return 0;
-}
 
 ecs_entity_t ecs_primitive_init(
     ecs_world_t *world,
